refactor(protocol): make map validation testable
