+++ client/src/include/map.h
@@ -86,15 +86,15 @@
 /**
  * The number of object layers.
  */
-#define NUM_LAYERS 7
+#define NUM_LAYERS MAP2_PROTOCOL_OBJECT_LAYERS
 /**
  * Number of sub-layers.
  */
-#define NUM_SUB_LAYERS 7
+#define NUM_SUB_LAYERS MAP2_PROTOCOL_SUB_LAYERS
 /**
  * Effective number of all the visible layers.
  */
-#define NUM_REAL_LAYERS (NUM_LAYERS * NUM_SUB_LAYERS)
+#define NUM_REAL_LAYERS MAP2_PROTOCOL_REAL_LAYERS
 
 #define GET_MAP_LAYER(_layer, _sub_layer) (NUM_LAYERS * (_sub_layer) + (_layer) - 1)
